refactor: Modularize PatientModelService into focused services
Refactored the original PatientModelService into three distinct services
to improve maintainability, testability, and adherence to file length guidelines:

1.  `PatientProfileService`: Handles CRUD operations for patient profiles
    and manages conversation history (`getProfileById`, `saveProfile`,
    `deleteProfile`, `getAvailableProfiles`, `addMessageToPatientHistory`).

2.  `BeliefConsistencyService`: Encapsulates the logic for checking
    belief consistency (`checkBeliefConsistency` and its internal helpers).

3.  `PatientResponseService`: Manages response generation and context
    creation (`createResponseContext`, `generatePatientPrompt`,
    `generateConsistentResponse`). This service now utilizes
    `PatientProfileService` and `BeliefConsistencyService`.

The original `PatientModelService.ts` has been deleted as its
responsibilities are now fully distributed.

Unit tests have been reorganized into separate files for each new service
(`PatientProfileService.test.ts`, `BeliefConsistencyService.test.ts`,
`PatientResponseService.test.ts`), and all tests are passing.
