+++ package.json
@@ -266,6 +266,7 @@
     "@testing-library/user-event": "^14.6.1",
     "@trunkio/launcher": "^1.3.4",
     "@types/archiver": "^6.0.3",
+    "@types/jest": "^30.0.0",
     "@types/k6": "^1.0.2",
     "@types/mdast": "^4.0.4",
     "@types/node": "^24.0.1",
